requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
pandas==2.1.3
python-dateutil==2.8.2
Pillow==10.1.0
whitenoise==6.6.0
gunicorn==21.2.0
